Remove all client helpers
