--- docs/api/client/extract.md
@@ -1,17 +1,0 @@
-# Extract
-
-::: matchbox.client.extract
-    options:
-        show_root_heading: true
-        show_root_full_path: true
-        members_order: source
-        show_if_no_docstring: true
-        docstring_style: google
-        show_signature_annotations: true
-        separate_signature: true
-        show_submodules: true
-        extra:
-            show_root_docstring: true
-        filters:
-            - "!^[A-Z]$"  # Excludes single-letter uppercase variables (like T, P, R)
-            - "!^_"       # Excludes private attributes
