+++ mkdocs.yml
@@ -10,7 +10,7 @@ nav:
 - Client:
   - Installation: client/install.md
   - Explore DAGs: client/explore-dags.md
-  - Retrieve matches: client/query-data.md
+  - Look up matches: client/look-up.md
   - Link and deduplicate: client/link-data.md
   - Evaluate models: client/evaluation.md
   - API: 
@@ -31,7 +31,6 @@ nav:
     - Models: api/client/models.md
     - Evaluation: api/client/eval.md
     - Results: api/client/results.md
-    - Extract: api/client/extract.md
   - Server: 
     - Overview: api/server/index.md
     - API: api/server/api.md
