+++ src/matchbox/client/_handler.py
@@ -227,7 +227,7 @@ def match(
     key: str,
     resolution: ResolutionName,
     threshold: int | None = None,
-) -> Match:
+) -> list[Match]:
     """Match a source against a list of targets."""
     log_prefix = f"Query {source}"
     logger.debug(
