+++ src/matchbox/client/dags.py
@@ -4,9 +4,13 @@ import datetime
 from collections import defaultdict
 from typing import Self
 
+from pyarrow import Table as ArrowTable
+
+from matchbox.client import _handler
 from matchbox.client.models import Model
 from matchbox.client.queries import Query
 from matchbox.client.sources import Source
+from matchbox.common.exceptions import MatchboxResolutionNotFoundError
 from matchbox.common.logging import logger
 
 
@@ -43,18 +47,19 @@ class DAG:
 
         self.nodes[step.name] = step
 
-    def _build_inverse_graph(self) -> tuple[dict[str, list[str]], str]:
-        """Build inverse graph and find the apex node.
+    @property
+    def final_step(self) -> str:
+        """Returns name of the root node in the DAG.
 
         Returns:
-            tuple: (inverse_graph, apex_node)
-
-                * inverse_graph: Dictionary mapping nodes to their parent nodes
-                * apex_node: The root node of the DAG
+            The name of the root node in the DAG
 
         Raises:
-            ValueError: If the DAG has multiple disconnected components
+            ValueError: If the DAG does not have a final step
         """
+        if not self.nodes:
+            raise ValueError("The DAG is empty.")
+
         inverse_graph = defaultdict(list)
         for node in self.graph:
             for neighbor in self.graph[node]:
@@ -66,7 +71,7 @@ class DAG:
         elif not apex_nodes:
             raise ValueError("No root node found, DAG might contain cycles")
         else:
-            return inverse_graph, apex_nodes.pop()
+            return apex_nodes.pop()
 
     def source(self, *args, **kwargs) -> Source:
         """Create Source and add it to the DAG."""
@@ -110,7 +115,7 @@ class DAG:
         Returns:
             String representation of the DAG with status indicators.
         """
-        _, root_name = self._build_inverse_graph()
+        root_name = self.final_step
         skipped = skipped or []
 
         def _get_node_status(name: str) -> str:
@@ -178,7 +183,7 @@ class DAG:
         start_time = datetime.datetime.now()
 
         # Determine order of execution steps
-        _, apex = self._build_inverse_graph()
+        root_node = self.final_step
 
         def depth_first(node: str, sequence: list):
             sequence.append(node)
@@ -187,7 +192,7 @@ class DAG:
                     depth_first(neighbour, sequence)
 
         inverse_sequence = []
-        depth_first(apex, inverse_sequence)
+        depth_first(root_node, inverse_sequence)
         sequence = list(reversed(inverse_sequence))
 
         # Identify skipped nodes
@@ -235,3 +240,110 @@ class DAG:
                 raise e
 
         logger.info("\n" + self.draw(start_time=start_time, skipped=skipped_nodes))
+
+    def lookup_key(
+        self,
+        from_source: str,
+        to_sources: list[str],
+        key: str,
+        threshold: int | None = None,
+    ) -> dict[str, list[str]]:
+        """Matches IDs against the selected backend.
+
+        Args:
+            from_source: Name of source the provided key belongs to
+            to_sources: Names of sources to find keys in
+            key: The value to match from the source. Usually a primary key
+            threshold (optional): The threshold to use for creating clusters.
+                If None, uses the resolutions' default threshold
+                If an integer, uses that threshold for the specified resolution, and the
+                resolution's cached thresholds for its ancestors
+
+        Examples:
+            ```python
+            dag.lookup_key(
+                from_source="companies_house",
+                to_sources=[
+                    "datahub_companies",
+                    "hmrc_exporters",
+                ]
+                key="8534735",
+            )
+            ```
+        """
+        matches = _handler.match(
+            targets=to_sources,
+            source=from_source,
+            key=key,
+            resolution=self.final_step,
+            threshold=threshold,
+        )
+
+        to_sources_results = {m.target: list(m.target_id) for m in matches}
+        # If no matches, _handler will raise
+        return {from_source: list(matches[0].source_id), **to_sources_results}
+
+    def extract_lookup(
+        self,
+        source_filter: list[str] | None = None,
+        location_names: list[str] | None = None,
+    ) -> ArrowTable:
+        """Return matchbox IDs to source key mapping, optionally filtering.
+
+        Args:
+            source_filter: An optional list of source resolution names to filter by.
+            location_names: An optional list of location names to filter by.
+        """
+        # Get all sources in scope of the resolution
+        point_of_truth = self.final_step
+        source_resolutions = _handler.get_leaf_source_resolutions(name=point_of_truth)
+
+        if source_filter:
+            source_resolutions = [
+                s for s in source_resolutions if s.name in source_filter
+            ]
+
+        if location_names:
+            source_resolutions = [
+                s
+                for s in source_resolutions
+                if s.config.location_config.name in location_names
+            ]
+
+        if not source_resolutions:
+            raise MatchboxResolutionNotFoundError("No compatible source was found")
+
+        source_mb_ids: list[ArrowTable] = []
+        source_to_key_field: dict[str, str] = {}
+
+        for s in source_resolutions:
+            # Get Matchbox IDs from backend
+            source_mb_ids.append(
+                _handler.query(
+                    source=s.name, resolution=point_of_truth, return_leaf_id=False
+                )
+            )
+
+            source_to_key_field[s.name] = s.config.key_field.name
+
+        # Join Matchbox IDs to form mapping table
+        mapping = source_mb_ids[0]
+        mapping = mapping.rename_columns(
+            {
+                "key": source_resolutions[0].config.qualified_key(
+                    source_resolutions[0].name
+                )
+            }
+        )
+        if len(source_resolutions) > 1:
+            for s, mb_ids in zip(
+                source_resolutions[1:], source_mb_ids[1:], strict=True
+            ):
+                mapping = mapping.join(
+                    right_table=mb_ids, keys="id", join_type="full outer"
+                )
+                mapping = mapping.rename_columns(
+                    {"key": s.config.qualified_key(s.name)}
+                )
+
+        return mapping
