+++ src/matchbox/common/factories/models.py
@@ -925,7 +925,7 @@ def query_to_model_factory(
         raise ValueError("Probabilities must be increasing values between 0 and 1")
 
     dag = left_query.dag
-    if right_query.dag != dag:
+    if right_query and right_query.dag != dag:
         raise ValueError("DAGs for left and right query must match.")
 
     # Check if right-side arguments are consistently provided
