--- test/client/helpers/test_match_helper.py
@@ -1,193 +1,0 @@
-import pytest
-from httpx import Response
-from respx import MockRouter
-from sqlalchemy import Engine
-
-from matchbox import match
-from matchbox.client.helpers.selector import Match
-from matchbox.common.dtos import BackendResourceType, NotFoundError
-from matchbox.common.exceptions import (
-    MatchboxEmptyServerResponse,
-    MatchboxResolutionNotFoundError,
-)
-from matchbox.common.factories.sources import source_factory
-
-
-def test_match_ok(matchbox_api: MockRouter, sqlite_warehouse: Engine):
-    """The client can perform the right call for matching."""
-    # Set up mocks
-    source_testkit = source_factory(
-        engine=sqlite_warehouse, name="source"
-    ).write_to_location()
-    target1_testkit = source_factory(
-        engine=sqlite_warehouse, name="target1"
-    ).write_to_location()
-    target2_testkit = source_factory(
-        engine=sqlite_warehouse, name="target2"
-    ).write_to_location()
-
-    mock_match1 = Match(
-        cluster=1,
-        source="source",
-        source_id={"a"},
-        target="target",
-        target_id={"b"},
-    )
-    mock_match2 = Match(
-        cluster=1,
-        source="source",
-        source_id={"a"},
-        target="target2",
-        target_id={"b"},
-    )
-    # The standard JSON serialiser does not handle Pydantic objects
-    serialised_matches = (
-        f"[{mock_match1.model_dump_json()}, {mock_match2.model_dump_json()}]"
-    )
-
-    match_route = matchbox_api.get("/match").mock(
-        return_value=Response(200, content=serialised_matches)
-    )
-    matchbox_api.get(f"/resolutions/{source_testkit.source.to_resolution().name}").mock(
-        return_value=Response(
-            200, json=source_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-    matchbox_api.get(
-        f"/resolutions/{target1_testkit.source.to_resolution().name}"
-    ).mock(
-        return_value=Response(
-            200, json=target1_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-    matchbox_api.get(
-        f"/resolutions/{target2_testkit.source.to_resolution().name}"
-    ).mock(
-        return_value=Response(
-            200, json=target2_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-
-    # Use match function
-    res = match(
-        "target1",
-        "target2",
-        source="source",
-        key="pk1",
-        resolution="foo",
-    )
-
-    # Verify results
-    assert len(res) == 2
-    assert isinstance(res[0], Match)
-    param_set = sorted(match_route.calls.last.request.url.params.multi_items())
-    assert param_set == sorted(
-        [
-            ("targets", "target1"),
-            ("targets", "target2"),
-            ("source", "source"),
-            ("key", "pk1"),
-            ("resolution", "foo"),
-        ]
-    )
-
-
-def test_match_404_resolution(matchbox_api: MockRouter, sqlite_warehouse: Engine):
-    """The client can handle a resolution not found error."""
-    # Set up mocks
-    source_testkit = source_factory(engine=sqlite_warehouse, name="source")
-    target_testkit = source_factory(engine=sqlite_warehouse, name="target")
-
-    matchbox_api.get("/match").mock(
-        return_value=Response(
-            404,
-            json=NotFoundError(
-                details="Resolution 42 not found",
-                entity=BackendResourceType.RESOLUTION,
-            ).model_dump(),
-        )
-    )
-    matchbox_api.get(f"/resolutions/{source_testkit.source.to_resolution().name}").mock(
-        return_value=Response(
-            200, json=source_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-    matchbox_api.get(f"/resolutions/{target_testkit.source.to_resolution().name}").mock(
-        return_value=Response(
-            200, json=target_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-
-    # Use match function
-    with pytest.raises(MatchboxResolutionNotFoundError, match="42"):
-        match(
-            "target",
-            source="source",
-            key="pk1",
-            resolution="foo",
-        )
-
-
-def test_match_404_source(matchbox_api: MockRouter, sqlite_warehouse: Engine):
-    """The client can handle a source not found error."""
-    target_testkit = source_factory(engine=sqlite_warehouse, name="target")
-
-    matchbox_api.get("/resolutions/nonexistent").mock(
-        return_value=Response(
-            404,
-            json=NotFoundError(
-                details="nonexistent", entity=BackendResourceType.RESOLUTION
-            ).model_dump(),
-        )
-    )
-    matchbox_api.get(f"/resolutions/{target_testkit.source.to_resolution().name}").mock(
-        return_value=Response(
-            200, json=target_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-
-    # Use match function
-    with pytest.raises(MatchboxResolutionNotFoundError, match="nonexistent"):
-        match(
-            "target",
-            source="nonexistent",
-            key="pk1",
-            resolution="foo",
-        )
-
-
-def test_match_empty_results_raises_exception(
-    matchbox_api: MockRouter, sqlite_warehouse: Engine
-):
-    """Test that match raises MatchboxEmptyServerResponse when no matches are found."""
-    # Set up mocks
-    source_testkit = source_factory(
-        engine=sqlite_warehouse, name="source"
-    ).write_to_location()
-    target_testkit = source_factory(
-        engine=sqlite_warehouse, name="target"
-    ).write_to_location()
-
-    # Mock empty match results
-    matchbox_api.get("/match").mock(return_value=Response(200, content="[]"))
-    matchbox_api.get(f"/resolutions/{source_testkit.source.to_resolution().name}").mock(
-        return_value=Response(
-            200, json=source_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-    matchbox_api.get(f"/resolutions/{target_testkit.source.to_resolution().name}").mock(
-        return_value=Response(
-            200, json=target_testkit.source.to_resolution().model_dump(mode="json")
-        )
-    )
-
-    # Test that empty match results raise MatchboxEmptyServerResponse
-    with pytest.raises(
-        MatchboxEmptyServerResponse, match="The match operation returned no data"
-    ):
-        match(
-            "target",
-            source="source",
-            key="pk1",
-            resolution="foo",
-        )
