+++ test/client/test_dags.py
@@ -1,15 +1,26 @@
 from datetime import datetime
 from unittest.mock import Mock, patch
 
+import polars as pl
+import pyarrow as pa
 import pytest
-from sqlalchemy import Engine
+from httpx import Response
+from polars.testing import assert_frame_equal
+from respx import MockRouter
+from sqlalchemy import Engine, create_engine
 
 from matchbox.client.dags import DAG
 from matchbox.client.models import Model
 from matchbox.client.models.dedupers import NaiveDeduper
 from matchbox.client.models.linkers import DeterministicLinker
 from matchbox.client.sources import Source
-from matchbox.common.factories.sources import source_factory
+from matchbox.common.arrow import SCHEMA_QUERY, table_to_buffer
+from matchbox.common.dtos import BackendResourceType, Match, NotFoundError
+from matchbox.common.exceptions import (
+    MatchboxEmptyServerResponse,
+    MatchboxResolutionNotFoundError,
+)
+from matchbox.common.factories.sources import source_factory, source_from_tuple
 
 
 @patch.object(Source, "run")
@@ -284,3 +295,296 @@ def test_dag_draw(sqlite_warehouse: Engine):
     assert all(
         indicator in tree_str_all_statuses for indicator in ["✅", "🔄", "⏸️", "⏭️"]
     )
+
+
+# Lookups
+
+
+def test_extract_lookup(sqlite_warehouse: Engine, matchbox_api: MockRouter):
+    """Entire lookup can be extracted from DAG."""
+    # Make dummy data
+    sqlite_memory_warehouse = create_engine("sqlite:///:memory:")
+
+    foo = source_from_tuple(
+        name="foo",
+        location_name="sqlite",
+        engine=sqlite_warehouse,
+        data_keys=[1, 2, 3],
+        data_tuple=({"col": 0}, {"col": 1}, {"col": 2}),
+    ).write_to_location()
+    bar = source_from_tuple(
+        name="bar",
+        location_name="sqlite_memory",
+        engine=sqlite_memory_warehouse,
+        data_keys=["a", "b", "c"],
+        data_tuple=({"col": 10}, {"col": 11}, {"col": 12}),
+    ).write_to_location()
+
+    dag = DAG("companies")
+    dag.source(**foo.into_dag()).query().linker(
+        dag.source(**bar.into_dag()).query(),
+        name="root",
+        model_class=DeterministicLinker,
+        model_settings={"left_id": "id", "right_id": "id", "comparisons": ""},
+    )
+
+    # Because of FULL OUTER JOIN, we expect some values to be null, and some explosions
+    expected_foo_bar_mapping = pl.DataFrame(
+        [
+            {"id": 1, "foo_key": "1", "bar_key": "a"},
+            {"id": 2, "foo_key": "2", "bar_key": None},
+            {"id": 3, "foo_key": "3", "bar_key": "b"},
+            {"id": 3, "foo_key": "3", "bar_key": "c"},
+        ]
+    )
+
+    # When selecting single source, we won't explode
+    expected_foo_mapping = expected_foo_bar_mapping.select(["id", "foo_key"]).unique()
+
+    # Mock API
+    matchbox_api.get("/resolutions/root/sources").mock(
+        return_value=Response(
+            200,
+            json=[
+                foo.source.to_resolution().model_dump(mode="json"),
+                bar.source.to_resolution().model_dump(mode="json"),
+            ],
+        )
+    )
+
+    matchbox_api.get("/query", params={"source": "foo"}).mock(
+        return_value=Response(
+            200,
+            content=table_to_buffer(
+                pa.Table.from_pylist(
+                    [
+                        {"id": 1, "key": "1"},
+                        {"id": 2, "key": "2"},
+                        {"id": 3, "key": "3"},
+                    ],
+                    schema=SCHEMA_QUERY,
+                )
+            ).read(),
+        )
+    )
+
+    matchbox_api.get("/query", params={"source": "bar"}).mock(
+        return_value=Response(
+            200,
+            content=table_to_buffer(
+                pa.Table.from_pylist(
+                    [
+                        {"id": 1, "key": "a"},
+                        {"id": 3, "key": "b"},
+                        {"id": 3, "key": "c"},
+                    ],
+                    schema=SCHEMA_QUERY,
+                )
+            ).read(),
+        )
+    )
+
+    # Case 0: No sources are found
+    with pytest.raises(MatchboxResolutionNotFoundError):
+        dag.extract_lookup(source_filter=["nonexistent"])
+
+    with pytest.raises(MatchboxResolutionNotFoundError):
+        dag.extract_lookup(location_names=["nonexistent"])
+
+    # Case 1: Retrieve single table
+    # With URI filter
+    foo_mapping = dag.extract_lookup(location_names=["sqlite"])
+
+    assert_frame_equal(
+        pl.from_arrow(foo_mapping),
+        expected_foo_mapping,
+        check_row_order=False,
+        check_column_order=False,
+    )
+
+    # With source filter
+    foo_mapping = dag.extract_lookup(source_filter="foo")
+
+    assert_frame_equal(
+        pl.from_arrow(foo_mapping),
+        expected_foo_mapping,
+        check_row_order=False,
+        check_column_order=False,
+    )
+
+    # With both filters
+    foo_mapping = dag.extract_lookup(source_filter="foo", location_names="sqlite")
+
+    assert_frame_equal(
+        pl.from_arrow(foo_mapping),
+        expected_foo_mapping,
+        check_row_order=False,
+        check_column_order=False,
+    )
+
+    # Case 2: Retrieve multiple tables
+    # With no filter
+    foo_bar_mapping = dag.extract_lookup()
+
+    assert_frame_equal(
+        pl.from_arrow(foo_bar_mapping),
+        expected_foo_bar_mapping,
+        check_row_order=False,
+        check_column_order=False,
+    )
+
+    # With source filter
+    foo_bar_mapping = dag.extract_lookup(source_filter=["foo", "bar"])
+
+    assert_frame_equal(
+        pl.from_arrow(foo_bar_mapping),
+        expected_foo_bar_mapping,
+        check_row_order=False,
+        check_column_order=False,
+    )
+
+
+def test_lookup_key_ok(matchbox_api: MockRouter, sqlite_warehouse: Engine):
+    """The DAG can map between single keys."""
+    # Set up dummy data
+    foo_testkit = source_factory(
+        engine=sqlite_warehouse, name="foo"
+    ).write_to_location()
+    bar_testkit = source_factory(
+        engine=sqlite_warehouse, name="bar"
+    ).write_to_location()
+    baz_testkit = source_factory(
+        engine=sqlite_warehouse, name="baz"
+    ).write_to_location()
+
+    dag = DAG("companies")
+    foo = dag.source(**foo_testkit.into_dag())
+    bar = dag.source(**bar_testkit.into_dag())
+    baz = dag.source(**baz_testkit.into_dag())
+
+    foo.query().linker(
+        bar.query(),
+        name="linker1",
+        model_class=DeterministicLinker,
+        model_settings={"left_id": "id", "right_id": "id", "comparisons": ""},
+    ).query(foo, bar).linker(
+        baz.query(),
+        name="linker2",
+        model_class=DeterministicLinker,
+        model_settings={"left_id": "id", "right_id": "id", "comparisons": ""},
+    )
+
+    mock_match1 = Match(
+        cluster=1, source="foo", source_id={"a"}, target="bar", target_id={"b"}
+    )
+    mock_match2 = Match(
+        cluster=1, source="foo", source_id={"a"}, target="baz", target_id={"b"}
+    )
+    # The standard JSON serialiser does not handle Pydantic objects
+    serialised_matches = (
+        f"[{mock_match1.model_dump_json()}, {mock_match2.model_dump_json()}]"
+    )
+
+    matchbox_api.get("/match").mock(
+        return_value=Response(200, content=serialised_matches)
+    )
+    matchbox_api.get(f"/resolutions/{foo.name}").mock(
+        return_value=Response(
+            200, json=foo_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+    matchbox_api.get(f"/resolutions/{bar.name}").mock(
+        return_value=Response(
+            200, json=bar_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+    matchbox_api.get(f"/resolutions/{baz.name}").mock(
+        return_value=Response(
+            200, json=baz_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+
+    # Use lookup function
+    matches = dag.lookup_key(from_source="foo", to_sources=["bar", "baz"], key="pk1")
+
+    # Verify results
+    assert matches == {foo.name: ["a"], bar.name: ["b"], baz.name: ["b"]}
+
+
+def test_lookup_key_404_source(matchbox_api: MockRouter):
+    """The client can handle a resolution not found error."""
+    # Set up dummy data
+    source_testkit = source_factory(name="source")
+    target_testkit = source_factory(name="target")
+
+    dag = DAG("companies")
+    dag.source(**source_testkit.into_dag()).query().linker(
+        dag.source(**target_testkit.into_dag()).query(),
+        name="root",
+        model_class=DeterministicLinker,
+        model_settings={"left_id": "id", "right_id": "id", "comparisons": ""},
+    )
+
+    matchbox_api.get("/match").mock(
+        return_value=Response(
+            404,
+            json=NotFoundError(
+                details="Resolution 42 not found",
+                entity=BackendResourceType.RESOLUTION,
+            ).model_dump(),
+        )
+    )
+    matchbox_api.get(f"/resolutions/{source_testkit.name}").mock(
+        return_value=Response(
+            200, json=source_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+    matchbox_api.get(f"/resolutions/{target_testkit.name}").mock(
+        return_value=Response(
+            200, json=target_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+
+    # Use match function
+    with pytest.raises(MatchboxResolutionNotFoundError, match="42"):
+        dag.lookup_key(from_source="source", to_sources=["target"], key="pk1")
+
+
+def test_match_empty_results_raises_exception(
+    matchbox_api: MockRouter, sqlite_warehouse: Engine
+):
+    """Test that match raises MatchboxEmptyServerResponse when no matches are found."""
+    # Set up dummy data
+    source_testkit = source_factory(
+        engine=sqlite_warehouse, name="source"
+    ).write_to_location()
+    target_testkit = source_factory(
+        engine=sqlite_warehouse, name="target"
+    ).write_to_location()
+
+    dag = DAG("companies")
+    dag.source(**source_testkit.into_dag()).query().linker(
+        dag.source(**target_testkit.into_dag()).query(),
+        name="root",
+        model_class=DeterministicLinker,
+        model_settings={"left_id": "id", "right_id": "id", "comparisons": ""},
+    )
+
+    # Mock empty match results
+    matchbox_api.get("/match").mock(return_value=Response(200, content="[]"))
+    matchbox_api.get(f"/resolutions/{source_testkit.source.to_resolution().name}").mock(
+        return_value=Response(
+            200, json=source_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+    matchbox_api.get(f"/resolutions/{target_testkit.source.to_resolution().name}").mock(
+        return_value=Response(
+            200, json=target_testkit.source.to_resolution().model_dump(mode="json")
+        )
+    )
+
+    # Test that empty match results raise MatchboxEmptyServerResponse
+    with pytest.raises(
+        MatchboxEmptyServerResponse, match="The match operation returned no data"
+    ):
+        dag.lookup_key(from_source="source", to_sources=["target"], key="pk1")
